refactor(test helpers): convert to TS
